backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Machine Learning
sentence-transformers==2.3.1
implicit==0.7.2
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS and HTTP
boto3==1.34.34
httpx==0.28.1
Pillow==10.2.0
google-generativeai>=0.3.0

# Data generation and testing
faker==22.6.0
pytest==7.4.4
pytest-asyncio==0.23.3
requests==2.32.3

# Additional utilities
python-dateutil>=2.8.2
PyJWT>=2.10.1
